Rename beyond_arena.py -> beyondarena_context.py
Make the contexts/ module names consistent with the classes they hold:
abstract_arena_context.py, tabarena_context.py, beyondarena_context.py.
Updates the lazy re-export map and all `tabarena.contexts.beyond_arena`
references (py + docs). Test file names are left as-is (renaming
test_beyond_arena.py would collide with test_beyond_arena_context.py); only
their imports change.
