+++ examples/!experimental/run_generate_beyondarena_core_subset.py
@@ -13,7 +13,7 @@ from pathlib import Path
 
 import pandas as pd
 
-from tabarena.contexts.beyond_arena import CORE_TASKS_CSV, BeyondArenaContext
+from tabarena.contexts.beyondarena_context import CORE_TASKS_CSV, BeyondArenaContext
 
 TARGET_RELIABILITY = 0.8
