+++ packages/tabarena/src/tabarena/contexts/__init__.py
@@ -16,7 +16,7 @@ from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
     from .abstract_arena_context import AbstractArenaContext
-    from .beyond_arena import BeyondArenaContext
+    from .beyondarena_context import BeyondArenaContext
     from .tabarena_context import TabArenaContext
 
 __all__ = ["AbstractArenaContext", "BeyondArenaContext", "TabArenaContext"]
@@ -24,7 +24,7 @@ __all__ = ["AbstractArenaContext", "BeyondArenaContext", "TabArenaContext"]
 _EXPORTS = {
     "AbstractArenaContext": "abstract_arena_context",
     "TabArenaContext": "tabarena_context",
-    "BeyondArenaContext": "beyond_arena",
+    "BeyondArenaContext": "beyondarena_context",
 }
