+++ packages/tabarena/src/tabarena/nips2025_utils/artifacts/_beyond_method_metadata.py
@@ -10,7 +10,7 @@ declared once in that model's ``models/<key>/info.py`` and reused for both the
 tabarena run and this Beyond re-run.
 
 ``beyond_method_metadata_collection`` is the aggregate consumed by
-:class:`~tabarena.contexts.beyond_arena.BeyondArenaContext`.
+:class:`~tabarena.contexts.beyondarena_context.BeyondArenaContext`.
 """
 
 from __future__ import annotations
