+++ packages/tabflow_slurm/experiments/run_setup_beyondarena_local.py
@@ -8,7 +8,7 @@ from pathlib import Path
 
 from tabarena.benchmark.experiment import BeyondArenaExperimentBundle
 from tabarena.benchmark.task.metadata import TaskSubset
-from tabarena.contexts.beyond_arena import BeyondArenaContext
+from tabarena.contexts.beyondarena_context import BeyondArenaContext
 from tabflow_slurm import (
     BeyondArenaResourcesSetup,
     LocalSequentialSetup,
