+++ tests/tabarena/evaluation/test_beyond_metadata.py
@@ -9,7 +9,7 @@ per-dataset frame.
 from __future__ import annotations
 
 from tabarena.benchmark.task.metadata import TaskMetadataCollection
-from tabarena.contexts.beyond_arena import BeyondArenaContext
+from tabarena.contexts.beyondarena_context import BeyondArenaContext
 from tabarena.evaluation.beyond_metadata import load_beyond_task_metadata_collection
